requirements.txt
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==38.0.4
Flask==2.2.2
Flask-MySQLdb==1.0.1
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.1.1
pycparser==2.21
PyJWT==2.6.0
Werkzeug==2.2.2
zipp==3.11.0
mysql-connector==2.2.9
mysql-connector-python==8.0.31
mysql-connector-python-rf==2.2.2